# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "list-it API",
    "description": "API for list-it UI",
    "version": "1.0.0"
  },
  "basePath": "/api/v1",
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "bearerToken": {
      "description": "OAuth2 bearer token generated using set options",
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://dev-1pzb3iuk.us.auth0.com/authorize",
      "tokenUrl": "https://dev-1pzb3iuk.us.auth0.com/oauth/token",
      "scopes": {
        "lst:user:read": "Grants read access",
        "lst:user:write": "Grants write access"
      },
      "name": "Authorization",
      "scheme": "bearer",
      "in": "header"
    }
  },
  "statusMessages": {
    "401_message": "Unauthorized!!!",
    "404_user_message": "User does not exist",
    "400_message": "Bad Request",
    "500_note_message": "Failed to create Note",
    "500_notes_list_message": "Failed to retrieve notes",
    "500_note_update_message": "Failed to update note",
    "500_message": "Internal server error"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check status.",
        "description": "To check and ensure API is running fine.",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": "application/json",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": {
              "$ref": "#/statusMessages/401_message"
            }
          }
        }
      }
    },
    "/getUserById": {
      "get": {
        "summary": "Get User details by user_id",
        "description": "To get User details by user_id",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": "application/json",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "description": "User ID of the user to fetch details for.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": {
              "$ref": "#/statusMessages/400_message"
            }
          },
          "401": {
            "description": {
              "$ref": "#/statusMessages/401_message"
            }
          },
          "404": {
            "description": {
              "$ref": "#/statusMessages/404_user_message"
            }
          }
        }
      }
    },
    "/saveNote": {
      "post": {
        "summary": "Create a new note",
        "description": "Create a new note with title, content and owner",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": "application/json",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Data for creating a new note",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SaveNoteRequest",
              "schema": {
                "$ref": "#/definitions/Note"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note with NoteId:<> created successfully"
          },
          "400": {
            "description": {
              "$ref": "#/statusMessages/400_message"
            }
          },
          "401": {
            "description": {
              "$ref": "#/statusMessages/401_message"
            }
          },
          "500": {
            "description": {
              "$ref": "#/statusMessages/500_note_message"
            }
          }
        }
      }
    },
    "/notes": {
      "get": {
        "summary": "Get a list of notes",
        "description": "Get list of owned and shared notes",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": "application/json",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "User ID of the user to fetch notes for",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetNotesResponse"
            }
          },
          "400": {
            "description": {
              "$ref": "#/statusMessages/400_message"
            }
          },
          "401": {
            "description": {
              "$ref": "#/statusMessages/401_message"
            }
          },
          "500": {
            "description": {
              "$ref": "#/statusMessages/500_notes_list_message"
            }
          }
        }
      }
    },
    "/shareNote": {
      "put": {
        "summary": "Share a note with a user",
        "description": "Update an existing note and share it with the respective userId",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": "application/json",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "UserId to share the note to",
            "required": true,
            "type": "string"
          },
          {
            "name": "noteId",
            "in": "query",
            "description": "NoteId to update the shared owner for the respective note",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref":"#/definitions/Note"
            }
          },
          "400": {
            "description": {
              "$ref": "#/statusMessages/400_message"
            }
          },
          "401": {
            "description": {
              "$ref": "#/statusMessages/401_message"
            }
          },
          "500": {
            "description": {
              "$ref": "#/statusMessages/500_note_update_message"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "user_id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "format": "password"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "nickname": {
          "type": "string"
        }
      }
    },
    "Note": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "title": {
          "type": "string"
        },
        "noteId": {
          "type": "string"
        },
        "ownerId": {
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      }
    },
    "SaveNoteRequest": {
      "type": "object",
      "properties": {
        "ownerId": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      }
    },
    "GetNotesResponse": {
      "type": "object",
      "properties": {
        "owned": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Note"
          }
        },
        "shared": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Note"
          }
        }
      }
    }
  }
}